python_scripts/requirements.txt
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
cssselect2==0.7.0
docutils==0.20.1
et-xmlfile==1.1.0
fonttools==4.44.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
markdown2==2.4.10
Markdown2PDF==0.1.4
MarkupSafe==2.1.3
mdpdf==0.0.18
openpyxl==3.1.2
packaging==23.2
pdfkit==1.0.0
Pillow==10.1.0
pycparser==2.21
pydyf==0.8.0
Pygments==2.16.1
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pyodbc==5.0.1
pyphen==0.14.0
python-dotenv==1.0.0
PyYAML==6.0.1
reportlab==4.0.6
requests==2.31.0
rst2pdf==0.101
six==1.16.0
smartypants==2.0.1
tinycss2==1.2.1
urllib3==2.0.5
weasyprint==60.1
webencodings==0.5.1
zipp==3.17.0
zopfli==0.2.3
